chore: 删除不再需要的配置文件和资源 - 删除项目私有配置文件 project.private.config.json - 删除 sitemap.json 和 sitemap3.json 文件 - 删除 sitemap48.json 文件 - 删除静态资源 c.jpg 和 sswd.png - 新增中国节假日相关的 util/china-holidays.js 文件 - 修改 http.js 文件中的 baseUrl 和 socketUrl - 新增随机弹窗文本的 util/popupTexts.js 文件 - 新增假期日历加载功能的 util/rq.js 文件

+++ pages/indexApp/xkt/xkt.vue
@@ -101,18 +101,24 @@
 				obeyMuteSwitch:false
 			})
 			this.audioCtx =wx.createInnerAudioContext()
-			this.audioCtx.src = 'http://www.daidaibg.com/statics/audio/di.mp3';
+			this.audioCtx.src = 'https://www.gaobug.com/img/static/xcx/audio/di.mp3';
 			this.audioBackground =wx.createInnerAudioContext()
-			this.audioBackground.src = 'http://www.daidaibg.com/statics/audio/air-extractor-fan.mp3';
+			this.audioBackground.src = 'https://www.gaobug.com/img/static/xcx/audio/air-extractor-fan.mp3';
 			this.audioBackground.loop =true;
 			this.audioOpen =wx.createInnerAudioContext()
-			this.audioOpen.src = 'http://www.daidaibg.com/statics/audio/ac-work.mp3';
+			this.audioOpen.src = 'https://www.gaobug.com/img/static/xcx/audio/ac-work.mp3';
 			
 		},
 		onUnload(){
-			this.audioCtx.destroy()
-			this.audioBackground.destroy()
-			this.audioOpen.destroy()
+			if (this.audioCtx && typeof this.audioCtx.destroy === 'function') {
+				this.audioCtx.destroy()
+			}
+			if (this.audioBackground && typeof this.audioBackground.destroy === 'function') {
+				this.audioBackground.destroy()
+			}
+			if (this.audioOpen && typeof this.audioOpen.destroy === 'function') {
+				this.audioOpen.destroy()
+			}
 		},
 		methods: {
 			switchMode(type) {
